UPD: user_api can now delete users
