+++ openedx/core/djangoapps/user_api/legacy_urls.py
@@ -33,6 +33,10 @@ urlpatterns = [
         r'^v1/preferences/time_zones/$',
         user_api_views.CountryTimeZoneListView.as_view(),
     ),
+    url(
+        r'^delete_user$',
+        user_api_views.DeleteUserView.as_view(),
+    ),
 ]
 
 if settings.FEATURES.get('ENABLE_COMBINED_LOGIN_REGISTRATION'):
