+++ openedx/core/djangoapps/user_api/serializers.py
@@ -109,3 +109,9 @@ class SoftwareSecurePhotoVerificationSerializer(serializers.ModelSerializer):
     class Meta(object):
         fields = ('status', 'expiration_datetime', 'is_verified')
         model = SoftwareSecurePhotoVerification
+
+class DeleteUserSerializer(serializers.Serializer):
+    """
+    Serializer for receiving list of usernames to delete
+    """
+    users = serializers.ListField()
