+++ openedx/core/djangoapps/user_api/urls.py
@@ -11,6 +11,8 @@ from .preferences.views import PreferencesDetailView, PreferencesView
 from .verification_api.views import PhotoVerificationStatusView
 from .validation.views import RegistrationValidationView
 
+from .views import DeleteUserView
+
 ME = AccountViewSet.as_view({
     'get': 'get',
 })
@@ -25,6 +27,11 @@ ACCOUNT_DETAIL = AccountViewSet.as_view({
 })
 
 urlpatterns = [
+    url (
+        r'^/delete_user$',
+        DeleteUserView.as_view(),
+        name = 'delete_user'
+    ),
     url(
         r'^v1/me$',
         ME,
